+++ lib/de_rjs.rb
@@ -1,0 +1,12 @@
+require 'rails'
+require 'active_support'
+
+unless defined? JQUERY_VAR
+  JQUERY_VAR = '$'
+end
+
+module DeRjs
+end
+
+require 'de_rjs/jquery_generator'
+require 'de_rjs/rewriter'
